Sync internal changes (#69)
* fix: Fix controller vmodel reconcilation in non-S3 storage case

Motivation

The predictor controller contains logic to reconcile a subset of the Predictor's status with a corresponding vmodels status, even in the case that the spec is invalid and the vmodel hasn't been updated.

Unfortunately this logic didn't account for the case that a vmodel does not exist, where a NOT_FOUND vmodel status would be returned from a getVModelStatus call. This resulted in a nil pointer dereference segfault.

Modifications

- Appropriately handle the case where the obtained VModelStatusInfo struct represents a non-existent vmodel (has nil ActiveModelState field). This should correspond to the Pending model state.
- Add FVT for this case

Result

Don't crash anymore when S3 storage not specified.

* perf: Some controller code cleanup and optimization

Motivation

Noticed some opportunities to streamline code while working on other controller features.

Modifications

- Make use of one-liner if statements in a bunch of places
- Use index when iterating over slices of structs (avoids copying)
- Other minor tidying

Result

Cleaner/faster code