feat: Introduce "multiModel" field to SR spec (#89)
* feat: Introduce multiModel field to SR Spec

This adds a field `multiModel` to the ServingRuntime spec which denotes
if a given SR is compatible with ModelMesh or not. The default is false.
ServingRuntime controller logic was adjusted to account for this field.

An e2e test was added to test this field.

Example SRs in docs were updated to include this field.

Signed-off-by: Paul Van Eck <pvaneck@us.ibm.com>

* Fix e2e test

Signed-off-by: Paul Van Eck <pvaneck@us.ibm.com>

* Rename function

Signed-off-by: Paul Van Eck <pvaneck@us.ibm.com>