feat: Add autoSelect ServingRuntime logic
In each supported model format of a ServingRuntime, a user can specify
autoSelect as true to indicate that the given ServingRuntime can be used
for automatic placement of a model of that format.

Logic was added to have MM only consider SRs with model formats
containing autoSelect as true when finding compatible runtimes.

The out-of-the-box ServingRuntimes now have "autoSelect: true" on all
of their supported formats, so there is no behavior change for users.

Signed-off-by: Paul Van Eck <pvaneck@us.ibm.com>