fix: Avoid nil pointer deref in FVT cleanup logic
Motivation

When an individual FVT fails, others are aborted and run cleanup logic. Part of this is disconnecting the gRPC client but it might not have been created at this point. So we often see additional nil pointer dereference panics in the FVT logs  in these cases. This can be confusing and distract from the root cause error.

Modification

Add nil check in the FVTClient DisconnectFromModelServing() pointer receiver function.

Result

Clearer FVT failure logs

Signed-off-by: Nick Hill <nickhill@us.ibm.com>