feat: Add TotalCopies in Predictor Status spec (#142)
**Motivation** 

`TotalCopies` is a recently added field to the InferenceService status which tracks the number
of copies/replicas of the model. We should start using it.

**Modifications**

`TotalCopies` was added to the `PredictorStatus`, and the Predictor CRD was regenerated.
The InferenceService registry now propagates this field.

**Result**

`TotalCopies` is now available in the `Predictor` and `InferenceService` statuses. 

Signed-off-by: Paul Van Eck <pvaneck@us.ibm.com>