feat: Support for protocolVersion-based runtime matching
Motivation

A protocolVersions list was recently added to the common ServingRuntime CRD, so that this can be taken into account when selecting a suitable runtime for a predictors that specify a protocol version.

Modifications

- Update the existing controller logic that orchestrates the model-mesh type constraints configuration to take into account the ServingRuntime protocolVersions list in addition to the existing supportedModelFormats list
- Update unit tests accordingly

Result

The runtime placement behaviour of model-mesh based predictors works as expected when they include a protocol version.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>