Fix etcd endpoints when etcd is in different namespace
Fix etcd endpoints when etcd is in different namespace
as described in https://github.com/kserve/modelmesh-serving/issues/186

The default etcd connection endpoints in quick start and fvt yaml are
changed to be placceholder based, and will be replaced during the
install script run. The etcd secret propagation will no longer assume
that etcd resides in the controller namespace.

Signed-off-by: Chin Huang <chhuang@us.ibm.com>