feat: Make list of supported built-in runtime server types overridable
Motivation

Model-mesh serving makes use of a single adapter image to interface with a number of "built-in" model server types, currently comprising triton, mlserver, ovms.

Currently which kind of adapter to run is controlled by the entrypoint to the adapter container within the modelmesh pods, which is set by the controller based on the builtInAdapter.serverType field of the ServingRuntime CRD.

However this is currently validated against a hardcoded list of type strings meaning that any extensions to the runtime adapter image to support new kinds of model servers also require a code change and rebuild of the controller image.

Modifications

Move the list of supported built-in server types to a list in the global config.
Update controller validation logic and tests accordingly.

Result

The shared built-in runtime adapter can be extended to support new runtime types without requiring a code change to the controller.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>