chore: Fix error in `verify_doc_links.py` script (#244)
#### Motivation

In its current version the `verify_doc_links.py` script excludes any URLs that contain a colon `:` character.

This was meant to exclude any local URLs with a port like "`:8008`". Instead it excludes any URLs that start with "`http:`" or "`https:`". The "rookie" who created that script! 🙄 


#### Modifications

Remove the ":" from the list of patterns to identify URLs that should be excluded from the link verification.

#### Result

220 links found instead of 211 before.


Signed-off-by: Christian Kadner <ckadner@us.ibm.com>