oauth-proxy: raise CPU request and limit.
With a CPU limit of 100m, oauth-proxy seems unable to cope with the
load associated with its own liveness probes, leading to the model
mesh pods being restarted every so often once a certain number of
routes (inference services) are created.

Raise the CPU limit from 100m to 2.
Raise the CPU request from 100m to 0.5.

Related-to: https://github.com/opendatahub-io/modelmesh-serving/issues/62
Related-to: https://github.com/opendatahub-io/modelmesh-serving/issues/16
Signed-off-by: François Cami <fcami@redhat.com>