fix: set non-empty modelPath for http storageURI (#382)
#### Motivation

When using the `storageURI` form of an HTTP (non Azure blob) address to download a model, the `modelPath` needs to be an non-empty string. Before this change, the `storageURI: http://models.r.us/my-model.json` form would be equivalent to the following `storage` spec:
```
storage:
  type: http
  path: ''    # this being empty is problematic for later processing
  parameters:
    url: http://models.r.us/path/to/my-model.json
```

The http storage type is currently the only way to have a valid storage configuration with an empty `path` (mainly because it has a "url" parameter that could include the full path). That said, I'm not sure if we should make a `path` required for the HTTP storage type. In particular, if the `url` is just `http://models.r.us/`, there is no path portion.

Related: https://github.com/kserve/modelmesh-runtime-adapter/issues/41#issuecomment-1564909111

#### Modifications

Set `modelPath` to the URL's Path and set the `url` parameter to not have the URL Path.

#### Result

With these changes, the `storageURI` example above changes to have a `path` field:

```
storage:
  type: http
  path: path/to/my-model.json
  parameters:
    url: http://models.r.us/
```

Signed-off-by: Travis Johnson <tsjohnso@us.ibm.com>
Signed-off-by: Christian Kadner <ckadner@us.ibm.com>

Co-authored-by: Christian Kadner <ckadner@us.ibm.com>
Signed-off-by: Luis Delossantos <luisgd@ibm.com>